created booking post and get api according requirement also handled anchoring regarding delete operation on users and vehicle table

+++ src/modular/users/user.services.ts
@@ -1,4 +1,5 @@
 import { pool } from "../../config/db";
+import sendJson from "../../helper/sendJson";
 
 const getAllUsers = async () => {
   const result = await pool.query(`SELECT id,name,email,phone,role FROM users`);
@@ -7,8 +8,13 @@ const getAllUsers = async () => {
 
 const updateUser = async (
   id: string,
-  payload: { name?: string; email?: string; phone?: string; role?: string }
+  payload: { name?: string; email?: string; phone?: string;role?: string},
+  isAdmin : boolean
 ) => {
+
+  if(!isAdmin){
+    delete payload.role
+  }
 
   const keys = Object.keys(payload);
   const values = Object.values(payload);
@@ -18,9 +24,9 @@ const updateUser = async (
   }
 
   const setQueries = keys
-    .map((key, index) => `${key}=$${index + 1}`)
+    .map((key, index) => `${key}=$${index + 1}` )
     .join(", ");
-
+  
 
   const result = await pool.query(
     `UPDATE users SET ${setQueries} WHERE id=$${keys.length + 1} RETURNING id,name,email,phone,role`,
@@ -30,6 +36,10 @@ const updateUser = async (
 };
 
 const deleteUser = async (id:string) => {
+    const hasAnyBookings = await pool.query(`SELECT * FROM bookings WHERE customer_id=$1`,[id]);
+    if(hasAnyBookings.rows.length){
+      throw new Error("user have ongoing booking cannot be deleted"); 
+    }
     const result = await pool.query(`DELETE FROM users WHERE id=$1`,[id]);
     return result;
 }
